refactor: proposal to atomicSwap util and sagas
