+++ src/sagas/index.js
@@ -1,7 +1,7 @@
 import { all, fork } from 'redux-saga/effects';
 import { saga as walletSagas } from './wallet';
 import { saga as tokensSagas } from './tokens';
-import { saga as proposalsSagas } from './proposals';
+import { saga as proposalsSagas } from './atomicSwap';
 
 function* defaultSaga() {
   yield all([
